// src/app/core/admin/models/ProductTable.ts
export interface ProductTable {
  product_id: 'Product Id';
  product_title: 'Product Title';
  product_name: 'Product Name';
  product_description: 'Product Description';
  product_quantity: 'Product Quantity';
  product_price: 'Product Price';
  discount_price: 'Discount Price';
  sku: 'SKU';
  tax: 'Tax';
  shipping_info_length?: 'Shipping Info Length';
  shipping_info_width?: 'Shipping Info Width';
  shipping_info_height?: 'Shipping Info Height';
  shipping_info_weight?: 'Shipping Info Weight';
  return_days?: 'Return Days';
  config?: 'Config';
  hsn_code?: 'HSN Code';
  meta_title?: 'Meta Title';
  meta_keyword?: 'Meta Keyword';
  meta_description?: 'Meta Description';
  custom_tags?: 'Custom Tags';
  key_feature_1?: 'Key Feature 1';
  key_feature_2?: 'Key Feature 2';
  key_feature_3?: 'Key Feature 3';
  is_active?: 'Status';
  actions?: 'Actions';
}